requirements.txt
diffusers==0.21.4
transformers==4.34.0
Pillow==10.3.0
tqdm==4.66.3
pandas==2.2.2
matplotlib==3.9.0
accelerate==0.32.1
timm==1.0.8
clean-fid==0.1.35